pyproject.toml
[tool.poetry]
name = "py-dashboard"
version = "0.1.0"
description = "Dashboard with Flask and Plotly"
authors = ["Arun Khattri"]
license = "GPLv3"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9, <=3.12"
pandas = "^2.1.4"
numpy = "^1.26.2"
plotly = "^5.18.0"
flask = "^3.0.0"
altair = "^5.2.0"
vega-datasets = "^0.9.0"


[tool.poetry.group.dev.dependencies]
black = "^23.12.0"
isort = "^5.13.2"
flake8 = "^6.1.0"
pytest = "^7.4.3"
ipython = "^8.18.1"
mypy = "^1.7.1"
jupyterlab = "^4.0.9"
python-lsp-server = "^1.9.0"
pylint = "^3.0.3"
rope = "^1.11.0"
python-decouple = "^3.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
